Move NewButton component into Button, update tests
